Refactor namespaces and code styling consistency
The main change is a refactor of the style used for conditional blocks, method formatting, variable declaration, and namespace formatting in several files. This change makes the style consistent and easier to read and maintain. Furthermore, an unnecessary comment related to command-line utility in the CMakeLists.txt has been removed. The changes don't affect the functionality of the code, but should improve readability and ease of development.

Made sure all necessary includes are present

Optimized initialization of structs `constexpr`, `noexcept`, list initializers etc.

Contributes to CURA-10561