Refactor version and name management in application build
This commit removes template file generation for cmdline.h.jinja and adds direct definitions in CMakeLists.txt and related files. The changes are necessary to improve the handling of name and version management across the application. This change leads to easier maintenance by having one single point of truth for the application version and building it directly into the binaries during the compile time. App names and versions are now defined using compile definitions in CMakeLists.txt. The .gitignore file is also updated to stop ignoring cmdline.h as it is no longer auto-generated.

Contributes to CURA-10561