Refactor Python bindings setup into separate CMakeLists.txt
The Python bindings setup was moved from the main CMakeLists.txt file into a new separate CMakeLists.txt within the pyDulcificum directory. This was done to modularise the CMake files and make the sections easier to manage and understand. This approach also provides better isolation of the configuration of different components.

Contributes to CURA-10561