Modify code for better traversal of GCode AST
The AST structure has been fundamentally changed from a standard vector to a multimap to support a tree-like structure, which is more representative of the GCode's structure and improves its traversal. A DFS function has been added to traverse the modified AST. A root node is returned from the parser and is used as the starting point for the DFS. This root node and visited nodes are managed by shared pointers for memory safety. The G0 operator overload has now been finalized. Changes in code also necessitated changes in the creators - they now return shared pointers. Lastly, the BaseEntry class has been introduced as parent of Entry class for polymorphism.

Contributes to CURA-10561