Fix calling next Command-template in list(s).
A big oversight. Also turns out that it still wouldn't compile in MSVC if I did this, but fortunately if the lists are focussed on consolidating the 'easy' and 'hard' ones together it works (since the easy ones will compile, and there aren't enough hard ones to justify splitting them up further).

part of CURA-11365