Update Python bindings and fix CMake library path
This commit updates the Python bindings to use cpython/3.12.2 and pybind11/2.11.1.
Additionally, a modification has been made to the CMake list to ensure the Conan CPython library is found, by setting the 'LD_LIBRARY_PATH' environment variable.
Other minor changes have been made to the 'conanfile.py' to correct the 'Python_ROOT_DIR' variable, along with the removal of an unused import.

Contribute to CURA-11079